test(integration): move FeedbackTool tests into integration suite and mark as @pytest.mark.slow; keep only pure unit tests in test_agents.py
